+++ src/rpc/cluster/mod.rs
@@ -5,7 +5,6 @@ mod sync_key_generator;
 mod sync_partial_key;
 mod sync_partial_keys;
 
-// mod submit_decryption_key;
 mod submit_final_reveal;
 mod submit_partial_key;
 mod sync_decryption_key;
@@ -13,7 +12,6 @@ mod sync_decryption_key;
 pub use get_key_generator_list::*;
 pub use get_skde_params::*;
 pub use request_submit_partial_key::*;
-// pub use submit_decryption_key::*;
 pub use submit_final_reveal::*;
 pub use submit_partial_key::*;
 pub use sync_decryption_key::*;
