+++ src/rpc/mod.rs
@@ -1,5 +1,4 @@
 pub mod cluster;
-pub mod common;
 pub mod solver;
 
 pub mod external;
