+++ src/rpc/solver/mod.rs
@@ -1,4 +1,3 @@
-
 mod submit_decryption_key;
 
 use submit_decryption_key::*;
